% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild-class.R
\name{as_ild}
\alias{as_ild}
\title{Coerce to ILD object}
\usage{
as_ild(x)
}
\arguments{
\item{x}{A data frame or tibble that may already be ILD-shaped.}
}
\value{
An ILD tibble with class \code{tidyild_df} and \code{ild_tbl}.
}
\description{
If the object already has the required `.ild_*` columns and
attributes, validates and returns it (with \code{tidyild_df} and \code{ild_tbl} class if missing).
Otherwise errors.
}
