% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{broom_ild_lme}
\alias{broom_ild_lme}
\alias{tidy.ild_lme}
\alias{augment.ild_lme}
\title{Tidy and augment ild_lme fits with broom.mixed}
\usage{
tidy.ild_lme(x, ...)

augment.ild_lme(x, ...)
}
\arguments{
\item{x}{A fitted model from [ild_lme()].}

\item{...}{Passed to \code{broom.mixed::tidy()} or \code{broom.mixed::augment()}.}
}
\value{
Same as the corresponding broom.mixed method.
}
\description{
These S3 methods delegate to [broom.mixed::tidy()] and [broom.mixed::augment()]
on the underlying model object so that ild_lme fits work in tidy workflows.
Package \pkg{broom.mixed} must be attached (e.g. \code{library(broom.mixed)}).
}
