% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild_generics.R, R/ild_brms_s3.R,
%   R/ild_ctsem_s3.R, R/ild_kfas_augment.R
\name{ild_augment}
\alias{ild_augment}
\alias{ild_augment.lmerMod}
\alias{ild_augment.lme}
\alias{ild_augment.default}
\alias{ild_augment.brmsfit}
\alias{ild_augment.ild_fit_ctsem}
\alias{ild_augment.ild_fit_kfas}
\title{Augment observations with fitted values and residuals (S3 generic)}
\usage{
ild_augment(x, ...)

\method{ild_augment}{lmerMod}(x, ...)

\method{ild_augment}{lme}(x, ...)

\method{ild_augment}{default}(x, ...)

\method{ild_augment}{brmsfit}(x, summary = TRUE, ...)

\method{ild_augment}{ild_fit_ctsem}(x, ...)

\method{ild_augment}{ild_fit_kfas}(x, ...)
}
\arguments{
\item{x}{A fitted model from [ild_lme()].}

\item{...}{Passed to [augment_ild_model()].}

\item{summary}{Logical. For \code{brmsfit}, use posterior predictive summary
for fitted values (default \code{TRUE}).}
}
\description{
Dispatches to [augment_ild_model()] for \code{lmerMod} and \code{lme} objects from [ild_lme()],
or the \code{brmsfit} method for [ild_brms()]. All methods return tables conforming to
\code{\link{ild_augment_schema}}.
}
