% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild_decomposition.R
\name{ild_center_plot}
\alias{ild_center_plot}
\title{Standalone WP/BP centering plot}
\usage{
ild_center_plot(x, ...)
}
\arguments{
\item{x}{An ILD object (see [is_ild()]).}

\item{...}{Variables to plot (tidy-select). Must be numeric. Only the first is plotted.}
}
\value{
A ggplot object (WP vs BP density overlay for the first selected variable).
}
\description{
Shows within-person deviation and between-person (person mean) distribution
for selected variable(s). Uses the same plot as [ild_decomposition(..., plot = TRUE)].
Useful when you only want the visualization without the variance table.
}
\examples{
d <- ild_simulate(n_id = 10, n_obs_per = 8, seed = 1)
x <- ild_prepare(d, id = "id", time = "time")
ild_center_plot(x, y)
}
