% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild_plot.R
\name{ild_heatmap}
\alias{ild_heatmap}
\title{ILD heatmap (alias for ild_plot with type = "heatmap")}
\usage{
ild_heatmap(x, var = NULL, ...)
}
\arguments{
\item{x}{ILD object or fitted model (for heatmap, data are taken from ild_data if model).}

\item{var}{Variable to plot. If NULL, single data column is used.}

\item{...}{Passed to [ild_plot()] (e.g. id_var, time_var).}
}
\value{
A ggplot object.
}
\description{
Person x time heatmap of a variable. See [ild_plot()].
}
