% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild_msm_fit.R
\name{ild_msm_diagnose}
\alias{ild_msm_diagnose}
\title{Diagnose an \code{ild_msm_fit} result in one call}
\usage{
ild_msm_diagnose(object, ...)
}
\arguments{
\item{object}{Output from [ild_msm_fit()].}

\item{...}{Passed to [ild_diagnose()].}
}
\value{
An \code{\link{ild_diagnostics_bundle}}.
}
\description{
Convenience bridge to [ild_diagnose()] using \code{object$fit} and
\code{object$weights_data}.
}
