% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild_kfas_plot.R
\name{ild_plot_states}
\alias{ild_plot_states}
\title{Plot smoothed latent states (first state by default)}
\usage{
ild_plot_states(x, state_index = 1L, ...)
}
\arguments{
\item{x}{An object from [ild_kfas()].}

\item{state_index}{Integer; which column of \code{alphahat} to plot (default \code{1L}).}

\item{...}{Passed to [ggplot2::labs()].}
}
\value{
A \code{ggplot} object.
}
\description{
Plot smoothed latent states (first state by default)
}
