% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild_plot.R
\name{ild_spaghetti}
\alias{ild_spaghetti}
\title{ILD spaghetti / person trajectories (alias for ild_plot with type = "trajectory")}
\usage{
ild_spaghetti(x, var = NULL, ...)
}
\arguments{
\item{x}{ILD object or fitted model.}

\item{var}{Variable to plot. If NULL, single data column is used.}

\item{...}{Passed to [ild_plot()] (e.g. max_ids, seed, id_var, time_var).}
}
\value{
A ggplot object.
}
\description{
Line plot of variable over time, one line per person. See [ild_plot()].
}
