% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild_generics.R, R/ild_brms_s3.R,
%   R/ild_ctsem_s3.R, R/ild_kfas_tidiers.R
\name{ild_tidy}
\alias{ild_tidy}
\alias{ild_tidy.lmerMod}
\alias{ild_tidy.lme}
\alias{ild_tidy.default}
\alias{ild_tidy.brmsfit}
\alias{ild_tidy.ild_fit_ctsem}
\alias{ild_tidy.ild_fit_kfas}
\title{Tidy fixed effects from an ILD model fit (S3 generic)}
\usage{
ild_tidy(x, ...)

\method{ild_tidy}{lmerMod}(x, ...)

\method{ild_tidy}{lme}(x, ...)

\method{ild_tidy}{default}(x, ...)

\method{ild_tidy}{brmsfit}(x, intervals = TRUE, object = FALSE, ...)

\method{ild_tidy}{ild_fit_ctsem}(x, ...)

\method{ild_tidy}{ild_fit_kfas}(x, ...)
}
\arguments{
\item{x}{A fitted model from [ild_lme()].}

\item{...}{Passed to [tidy_ild_model()].}

\item{intervals}{Logical. For \code{brmsfit} from [ild_brms()], include
posterior \code{rhat}, \code{ess_bulk}, \code{ess_tail} (default \code{TRUE}).}

\item{object}{Logical. Passed through for non-brms fits; ignored for \code{brmsfit}.}
}
\description{
Dispatches to [tidy_ild_model()] for \code{lmerMod} and \code{lme} objects from [ild_lme()],
or the \code{brmsfit} method for [ild_brms()]. All methods return tables conforming to
\code{\link{ild_tidy_schema}}.
}
