% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ild_tsibble.R
\name{ild_tsibble_meta}
\alias{ild_tsibble_meta}
\title{Tsibble provenance from an ILD object}
\usage{
ild_tsibble_meta(x)
}
\arguments{
\item{x}{An object that passes [validate_ild()].}
}
\value{
A named list or \code{NULL}.
}
\description{
If [ild_prepare()] was called on a \code{tbl_ts}, metadata from the source
tsibble (key, index, interval summary, regularity) is stored in
\code{attr(x, "tidyILD")$tsibble}. Returns \code{NULL} if the object was
not prepared from a tsibble.
}
\seealso{
[ild_prepare()], [ild_as_tsibble()],
  \href{../doc/tsibble-interoperability.html}{\code{vignette("tsibble-interoperability", package = "tidyILD")}}
}
